Renamed folders. Updates to SingletonScriptableObjects. Updates to SettingsDatabase.

+++ Runtime/Settings/ISettingsDatabase.cs
@@ -2,6 +2,6 @@ namespace Kryz.Settings
 {
 	public interface ISettingsDatabase
 	{
-		public T Get<T>(uint id) where T : SettingsAsset;
+		public T Get<T>(ulong id) where T : SettingsAsset;
 	}
 }
